OK. Major break-through: this draws four lines and they rotate.
I think there may be a bug somewhere because bochs plays up, but qemu doesn't, and either there's a bug in bochs, or there is a bug in my code that qemu is masking. Money's on a bug in my code. ANYWAY. This is HUGE.

Oh, this also broke the two-sectors limits and we are now using three because bytes are incompressible and I'm not smart enough.